+++ src/templates.js
@@ -254,51 +254,6 @@ function trending(subreddits) {
   return html;
 }
 
-// panel to display empty subreddit error
-function empty(subreddit) {
-  return `
-  <p>
-    <span class="keyword-color">function </span>
-    <span class="function-color">loadArticles</span><span class="bracket-color">(</span>
-    <span class="variable-color">id </span><span class="bracket-color">) {</span>
-    
-    </br>
-    
-    <span class="keyword-color">&nbsp; &nbsp; try</span>
-    <span class="bracket-color">{</span>
-    
-    </br>
-    
-    <span class="variable-color">&nbsp; &nbsp; &nbsp; &nbsp; api.</span><span class="function-color">get</span><span class="bracket-color">(</span>
-    <span class="string-color">'/articles/'</span>
-    <span class="variable-color">+ id</span><span class="bracket-color">)</span><span class="variable-color">;</span>
-    
-    </br>
-    
-    <span class="bracket-color">&nbsp; &nbsp; }</span>
-    
-    </br>
-    
-    <span class="keyword-color">&nbsp; &nbsp; catch</span><span class="bracket-color">(</span>
-    <span class="variable-color">err</span>
-    <span class="bracket-color">) {</span>
-    
-    </br>
-    
-    <span class="variable-color">&nbsp; &nbsp; &nbsp; &nbsp; window.</span><span class="function-color">alert</span><span class="bracket-color">(</span>
-    <span class="string-color">'Subreddit "${subreddit}" seems to be empty or not available'</span><span class="variable-color">, err</span><span class="bracket-color">)</span><span class="variable-color">;</span>
-    
-    </br>
-    
-    <span class="bracket-color">&nbsp; &nbsp; }</span>
-    
-    </br>
-    
-    <span class="bracket-color">}</span>
-  </p>
-  `;
-}
-
 // panel to display subreddit search
 function search(subreddit) {
   return `